# arena-app/api/generated/models/createContactDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Arena API
 * Arena Messenger Backend API
 * OpenAPI spec version: 1.0
 */

export interface CreateContactDto {
  /** 추가할 유저 ID */
  userId: string;
}