# arena-app/api/generated/models/createUserDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Arena API
 * Arena Messenger Backend API
 * OpenAPI spec version: 1.0
 */

export interface CreateUserDto {
  /**
   * @minLength 1
   * @maxLength 32
   */
  nick: string;
  email?: string;
  /** @maxLength 140 */
  statusMessage?: string;
}