# arena-app/api/generated/models/messageDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Arena API
 * Arena Messenger Backend API
 * OpenAPI spec version: 1.0
 */
import type { UserDto } from './userDto';

export interface MessageDto {
  /** 메시지 ID */
  messageId: string;
  /** 채널 ID */
  channelId: string;
  /** 보낸 사람 */
  sender: UserDto;
  /** 메시지 순번 */
  seq: number;
  /** 메시지 내용 */
  content: string;
  /** 생성 시간 */
  createdAt: string;
  /** 수정 시간 */
  updatedAt: string;
}