# arena-app/api/generated/models/registerDeviceDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Arena API
 * Arena Messenger Backend API
 * OpenAPI spec version: 1.0
 */
import type { RegisterDeviceDtoPlatform } from './registerDeviceDtoPlatform';

export interface RegisterDeviceDto {
  deviceId: string;
  fcmToken: string;
  platform: RegisterDeviceDtoPlatform;
  deviceModel?: string;
  osVersion?: string;
}