# arena-app/api/generated/models/userDto.ts
/**
 * Generated by orval v7.17.0 🍺
 * Do not edit manually.
 * Arena API
 * Arena Messenger Backend API
 * OpenAPI spec version: 1.0
 */
import type { UserDtoAvatar } from './userDtoAvatar';

export interface UserDto {
  /** 유저 ID */
  userId: string;
  /** 유저 태그 */
  utag: string;
  /** 닉네임 */
  nick: string;
  /**
   * 프로필 이미지
   * @nullable
   */
  avatar?: UserDtoAvatar;
  /**
   * 이메일
   * @nullable
   */
  email?: string | null;
  /**
   * 상태 메시지
   * @nullable
   */
  statusMessage?: string | null;
  /** 생성 시간 */
  createdAt: string;
  /** 수정 시간 */
  updatedAt: string;
}