# arena-app/api/swagger.json
{"openapi":"3.0.0","paths":{"/api/v1/users/me":{"get":{"operationId":"getMe","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto_Nullable_Result"}}}}},"security":[{"bearer":[]}],"summary":"내 정보 조회","tags":["users"]},"patch":{"operationId":"updateMe","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UpdateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto_Result"}}}}},"security":[{"bearer":[]}],"summary":"내 정보 수정","tags":["users"]}},"/api/v1/users/{userId}":{"get":{"operationId":"getUser","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto_Result"}}}}},"security":[{"bearer":[]}],"summary":"유저 조회","tags":["users"]}},"/api/v1/users":{"post":{"operationId":"createUser","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateUserDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"회원가입","tags":["users"]}},"/api/v1/files/presigned-url":{"post":{"operationId":"getPresignedUrl","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPresignedUrlDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresignedUrlDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"파일 업로드 URL 발급","tags":["files"]}},"/api/v1/files/register":{"post":{"operationId":"registerFile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterFileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"파일 등록","tags":["files"]}},"/api/v1/files/private/presigned-url":{"post":{"operationId":"getPrivatePresignedUrl","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GetPresignedUrlDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PresignedUrlDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Private 파일 업로드 URL 발급","tags":["files"]}},"/api/v1/files/private/register":{"post":{"operationId":"registerPrivateFile","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterFileDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"Private 파일 등록","tags":["files"]}},"/api/v1/files/{fileId}":{"get":{"operationId":"getFile","parameters":[{"name":"fileId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/FileDto_Result"}}}}},"security":[{"bearer":[]}],"summary":"파일 조회","tags":["files"]},"delete":{"operationId":"deleteFile","parameters":[{"name":"fileId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResultDto"}}}}},"security":[{"bearer":[]}],"summary":"파일 삭제","tags":["files"]}},"/api/v1/devices/register":{"post":{"operationId":"registerDevice","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/RegisterDeviceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResultDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResultDto"}}}}},"security":[{"bearer":[]}],"summary":"디바이스 등록","tags":["devices"]}},"/api/v1/devices/unregister":{"post":{"operationId":"unregisterDevice","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UnregisterDeviceDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResultDto"}}}},"201":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResultDto"}}}}},"security":[{"bearer":[]}],"summary":"디바이스 해제","tags":["devices"]}},"/api/v1/channels/direct":{"post":{"operationId":"createDirectChannel","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateDirectChannelDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"DM 생성 (이미 있으면 기존 반환)","tags":["channels"]}},"/api/v1/channels/group":{"post":{"operationId":"createGroupChannel","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateGroupChannelDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"그룹 채널 생성","tags":["channels"]}},"/api/v1/channels":{"get":{"operationId":"getMyChannels","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelDto_ListResult"}}}}},"security":[{"bearer":[]}],"summary":"내 채널 목록","tags":["channels"]}},"/api/v1/channels/{channelId}":{"get":{"operationId":"getChannel","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ChannelDto_Result"}}}}},"security":[{"bearer":[]}],"summary":"채널 상세 조회","tags":["channels"]}},"/api/v1/messages/channel/{channelId}":{"post":{"operationId":"createMessage","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateMessageDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"메시지 보내기","tags":["messages"]},"get":{"operationId":"getMessages","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"before","required":false,"in":"query","description":"이 메시지 이전 (이 메시지 미포함)","schema":{"type":"string"}},{"name":"after","required":false,"in":"query","description":"이 메시지 이후 (이 메시지 미포함)","schema":{"type":"string"}},{"name":"around","required":false,"in":"query","description":"이 메시지 기준 앞뒤로 조회 (이 메시지 포함)","schema":{"type":"string"}},{"name":"limit","required":false,"in":"query","description":"가져올 개수 (기본 50, 최대 100)","schema":{"minimum":1,"maximum":100,"type":"number"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageDto_InfinityListResult"}}}}},"security":[{"bearer":[]}],"summary":"메시지 목록 조회","tags":["messages"]}},"/api/v1/messages/channel/{channelId}/sync":{"get":{"operationId":"syncMessages","parameters":[{"name":"channelId","required":true,"in":"path","schema":{"type":"string"}},{"name":"since","required":true,"in":"query","description":"이 시간 이후 변경된 메시지 조회 (ISO 8601)","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageSyncDataDto_Result"}}}}},"security":[{"bearer":[]}],"summary":"메시지 동기화","tags":["messages"]}},"/api/v1/contacts":{"get":{"operationId":"getContacts","parameters":[],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactDto_ListResult"}}}}},"security":[{"bearer":[]}],"summary":"내 연락처 목록 조회","tags":["contacts"]},"post":{"operationId":"createContact","parameters":[],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateContactDto"}}}},"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ContactDto_Result"}}}},"201":{"description":""}},"security":[{"bearer":[]}],"summary":"연락처 추가","tags":["contacts"]}},"/api/v1/contacts/{userId}":{"delete":{"operationId":"deleteContact","parameters":[{"name":"userId","required":true,"in":"path","schema":{"type":"string"}}],"responses":{"200":{"description":"","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ApiResultDto"}}}}},"security":[{"bearer":[]}],"summary":"연락처 삭제","tags":["contacts"]}}},"info":{"title":"Arena API","description":"Arena Messenger Backend API","version":"1.0","contact":{}},"tags":[{"name":"arena","description":""}],"servers":[],"components":{"securitySchemes":{"bearer":{"scheme":"bearer","bearerFormat":"JWT","type":"http","description":"Enter your JWT token"}},"schemas":{"FileDto":{"type":"object","properties":{"fileId":{"type":"string","description":"파일 ID"},"url":{"type":"string","description":"파일 URL"},"name":{"type":"string","description":"파일명"},"mimeType":{"type":"string","description":"MIME 타입"},"size":{"type":"number","description":"파일 크기 (bytes)"},"createdAt":{"format":"date-time","type":"string","description":"생성 시간"},"updatedAt":{"format":"date-time","type":"string","description":"수정 시간"}},"required":["fileId","url","name","mimeType","size","createdAt","updatedAt"]},"UserDto":{"type":"object","properties":{"userId":{"type":"string","description":"유저 ID"},"utag":{"type":"string","description":"유저 태그"},"nick":{"type":"string","description":"닉네임"},"avatar":{"nullable":true,"description":"프로필 이미지","allOf":[{"$ref":"#/components/schemas/FileDto"}]},"email":{"type":"string","nullable":true,"description":"이메일"},"statusMessage":{"type":"string","nullable":true,"description":"상태 메시지"},"createdAt":{"format":"date-time","type":"string","description":"생성 시간"},"updatedAt":{"format":"date-time","type":"string","description":"수정 시간"}},"required":["userId","utag","nick","createdAt","updatedAt"]},"UserDto_Nullable_Result":{"type":"object","properties":{"data":{"nullable":true,"allOf":[{"$ref":"#/components/schemas/UserDto"}]},"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true}},"required":["data","success","errorCode"]},"UserDto_Result":{"type":"object","properties":{"data":{"nullable":false,"allOf":[{"$ref":"#/components/schemas/UserDto"}]},"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true}},"required":["data","success","errorCode"]},"UpdateUserDto":{"type":"object","properties":{"nick":{"type":"string","minLength":1,"maxLength":32},"statusMessage":{"type":"string","maxLength":140},"avatarFileId":{"type":"string","nullable":true,"description":"아바타 파일 ID (null이면 아바타 제거)"}}},"CreateUserDto":{"type":"object","properties":{"nick":{"type":"string","minLength":1,"maxLength":32},"email":{"type":"string","format":"email"},"statusMessage":{"type":"string","maxLength":140}},"required":["nick"]},"GetPresignedUrlDto":{"type":"object","properties":{"directory":{"type":"string","description":"디렉토리 (예: avatars, attachments, group-icons)","example":"avatars"},"mimeType":{"type":"string","description":"MIME 타입","example":"image/jpeg"}},"required":["mimeType"]},"PresignedUrlDto":{"type":"object","properties":{"uploadUrl":{"type":"string","description":"S3 업로드 URL"},"key":{"type":"string","description":"S3 key (업로드 후 서버에 전달)"},"expiresIn":{"type":"number","description":"URL 만료 시간 (초)"}},"required":["uploadUrl","key","expiresIn"]},"PresignedUrlDto_Result":{"type":"object","properties":{"data":{"nullable":false,"allOf":[{"$ref":"#/components/schemas/PresignedUrlDto"}]},"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true}},"required":["data","success","errorCode"]},"RegisterFileDto":{"type":"object","properties":{"key":{"type":"string","description":"Storage key (presigned URL 발급 시 받은 값)"},"name":{"type":"string","description":"원본 파일명"}},"required":["key","name"]},"FileDto_Result":{"type":"object","properties":{"data":{"nullable":false,"allOf":[{"$ref":"#/components/schemas/FileDto"}]},"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true}},"required":["data","success","errorCode"]},"ApiResultDto":{"type":"object","properties":{"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true}},"required":["success","errorCode"]},"RegisterDeviceDto":{"type":"object","properties":{"deviceId":{"type":"string"},"fcmToken":{"type":"string"},"platform":{"type":"string","enum":["ios","android","web"]},"deviceModel":{"type":"string"},"osVersion":{"type":"string"}},"required":["deviceId","fcmToken","platform"]},"UnregisterDeviceDto":{"type":"object","properties":{"deviceId":{"type":"string"}},"required":["deviceId"]},"CreateDirectChannelDto":{"type":"object","properties":{"userId":{"type":"string","description":"대화 상대 userId"}},"required":["userId"]},"ParticipantDto":{"type":"object","properties":{"user":{"description":"유저 정보","allOf":[{"$ref":"#/components/schemas/UserDto"}]},"lastReadAt":{"format":"date-time","type":"string","nullable":true,"description":"마지막 읽은 시간"},"createdAt":{"format":"date-time","type":"string","description":"참여 시간"},"updatedAt":{"format":"date-time","type":"string","description":"수정 시간"}},"required":["user","createdAt","updatedAt"]},"ChannelDto":{"type":"object","properties":{"channelId":{"type":"string","description":"채널 ID"},"type":{"type":"string","description":"채널 타입","enum":["direct","group","team"]},"name":{"type":"string","nullable":true,"description":"채널 이름 (DM은 null)"},"icon":{"nullable":true,"description":"채널 아이콘 (DM은 null)","allOf":[{"$ref":"#/components/schemas/FileDto"}]},"participants":{"description":"참여자 목록","type":"array","items":{"$ref":"#/components/schemas/ParticipantDto"}},"lastMessageAt":{"format":"date-time","type":"string","nullable":true,"description":"마지막 메시지 시간"},"createdAt":{"format":"date-time","type":"string","description":"생성 시간"},"updatedAt":{"format":"date-time","type":"string","description":"수정 시간"}},"required":["channelId","type","participants","createdAt","updatedAt"]},"ChannelDto_Result":{"type":"object","properties":{"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true},"data":{"nullable":false,"allOf":[{"$ref":"#/components/schemas/ChannelDto"}]}},"required":["success","errorCode","data"]},"CreateGroupChannelDto":{"type":"object","properties":{"name":{"type":"string","maxLength":100,"description":"그룹 이름"},"userIds":{"description":"초대할 유저 ID 목록","type":"array","items":{"type":"string"}},"iconFileId":{"type":"string","description":"그룹 아이콘 파일 ID"}},"required":["name"]},"ChannelDto_ListResult":{"type":"object","properties":{"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true},"data":{"nullable":false,"type":"array","items":{"$ref":"#/components/schemas/ChannelDto"}}},"required":["success","errorCode","data"]},"CreateMessageDto":{"type":"object","properties":{"content":{"type":"string","description":"메시지 내용"}},"required":["content"]},"MessageDto":{"type":"object","properties":{"messageId":{"type":"string","description":"메시지 ID"},"channelId":{"type":"string","description":"채널 ID"},"sender":{"description":"보낸 사람","allOf":[{"$ref":"#/components/schemas/UserDto"}]},"seq":{"type":"number","description":"메시지 순번"},"content":{"type":"string","description":"메시지 내용"},"createdAt":{"format":"date-time","type":"string","description":"생성 시간"},"updatedAt":{"format":"date-time","type":"string","description":"수정 시간"}},"required":["messageId","channelId","sender","seq","content","createdAt","updatedAt"]},"MessageDto_Result":{"type":"object","properties":{"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true},"data":{"nullable":false,"allOf":[{"$ref":"#/components/schemas/MessageDto"}]}},"required":["success","errorCode","data"]},"MessageDto_InfinityListResult":{"type":"object","properties":{"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true},"data":{"type":"array","items":{"$ref":"#/components/schemas/MessageDto"}},"hasNext":{"type":"boolean","description":"다음 데이터 존재 여부"},"hasPrev":{"type":"boolean","description":"이전 데이터 존재 여부"}},"required":["success","errorCode","data","hasNext","hasPrev"]},"MessageSyncDataDto":{"type":"object","properties":{"created":{"description":"새로 생성된 메시지","type":"array","items":{"$ref":"#/components/schemas/MessageDto"}},"updated":{"description":"수정된 메시지","type":"array","items":{"$ref":"#/components/schemas/MessageDto"}},"deleted":{"description":"삭제된 메시지 ID","type":"array","items":{"type":"string"}},"requireFullSync":{"type":"boolean","description":"true면 전체 재조회 필요"}},"required":["created","updated","deleted","requireFullSync"]},"MessageSyncDataDto_Result":{"type":"object","properties":{"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true},"data":{"nullable":false,"allOf":[{"$ref":"#/components/schemas/MessageSyncDataDto"}]}},"required":["success","errorCode","data"]},"ContactDto":{"type":"object","properties":{"user":{"description":"연락처에 저장된 유저 정보","allOf":[{"$ref":"#/components/schemas/UserDto"}]},"createdAt":{"format":"date-time","type":"string","description":"추가한 시간"},"updatedAt":{"format":"date-time","type":"string","description":"수정 시간"}},"required":["user","createdAt","updatedAt"]},"ContactDto_ListResult":{"type":"object","properties":{"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true},"data":{"nullable":false,"type":"array","items":{"$ref":"#/components/schemas/ContactDto"}}},"required":["success","errorCode","data"]},"CreateContactDto":{"type":"object","properties":{"userId":{"type":"string","description":"추가할 유저 ID"}},"required":["userId"]},"ContactDto_Result":{"type":"object","properties":{"success":{"type":"boolean"},"errorCode":{"type":"string","nullable":true},"data":{"nullable":false,"allOf":[{"$ref":"#/components/schemas/ContactDto"}]}},"required":["success","errorCode","data"]}}}}